+++ CHANGES.md
@@ -6,6 +6,7 @@
 * changed to use dashicon to identify private repos in Settings
 * fixed transient update when doing shiny updates
 * moved action/filter hook calls completely out of constructors, make @carlalexander happy
+* added ability to update from GitHub release asset
 
 #### 6.0.0 / 2016-10-26
 * added `class Language_Pack` and new repo, [Language Pack Maker](https://github.com/afragen/github-updater-language-pack-maker), to create and update from a separate Language Pack repository.